pyproject.toml
[project]
name = "vibe-coding-homework-1"
version = "0.1.0"
description = "Vibe Coding - Lekce 1: Python skript pro LLM API s nastrojem (tool use)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "python-dotenv>=1.0.1",
]
